+++ packages/use-intl/dts.config.js
@@ -5,9 +5,9 @@
  */
 module.exports = {
   rollup(config) {
-    if (config.output.format === 'esm') {
-      config.output.preserveModules = true;
-    }
+    // if (config.output.format === 'esm') {
+    //   config.output.preserveModules = true;
+    // }
     return config;
   }
 };
