+++ packages/use-intl/src/IntlProvider.tsx
@@ -2,9 +2,8 @@ import React, {ReactNode, useEffect} from 'react';
 import AbstractIntlMessages from './AbstractIntlMessages';
 import Formats from './Formats';
 import IntlContext from './IntlContext';
+import IntlError, {IntlErrorCode} from './IntlError';
 import {RichTranslationValues} from './TranslationValues';
-import validateMessages from './validateMessages';
-import {IntlError} from '.';
 
 type Props = {
   /** All messages that will be available in your components. */
@@ -68,12 +67,52 @@ export default function IntlProvider({
   messages,
   ...contextValues
 }: Props) {
+  // Validation of the namespace keys is currently inlined as the bundler that
+  // Bundlephobia uses is apparently not smart enough to throw the import out if
+  // this is imported from a separate module.
   if (__DEV__) {
     // eslint-disable-next-line react-hooks/rules-of-hooks
     useEffect(() => {
-      if (messages) {
-        validateMessages(messages, onError);
+      function validateMessagesSegment(
+        messagesPart: AbstractIntlMessages,
+        invalidKeyLabels: Array<string>,
+        parentPath?: string
+      ) {
+        Object.entries(messagesPart).forEach(([key, messageOrMessages]) => {
+          if (key.includes('.')) {
+            let keyLabel = key;
+            if (parentPath) keyLabel += ` (at ${parentPath})`;
+            invalidKeyLabels.push(keyLabel);
+          }
+          if (typeof messageOrMessages === 'object') {
+            validateMessagesSegment(
+              messageOrMessages,
+              invalidKeyLabels,
+              [parentPath, key].filter((part) => part != null).join('.')
+            );
+          }
+        });
       }
+
+      function validateMessages() {
+        if (!messages) return;
+
+        const invalidKeyLabels: Array<string> = [];
+        validateMessagesSegment(messages, invalidKeyLabels);
+
+        if (invalidKeyLabels.length > 0) {
+          onError(
+            new IntlError(
+              IntlErrorCode.INVALID_KEY,
+              `Namespace keys can not contain the character "." as this is used to express nesting. Please remove it or replace it with another character.\n\nInvalid ${
+                invalidKeyLabels.length === 1 ? 'key' : 'keys'
+              }: ${invalidKeyLabels.join(', ')}`
+            )
+          );
+        }
+      }
+
+      validateMessages();
     }, [messages, onError]);
   }
